+++ TablePro/Models/Settings/EditorSettings.swift
@@ -76,7 +76,6 @@ struct EditorSettings: Codable, Equatable {
     var uppercaseKeywords: Bool
     var queryParametersEnabled: Bool
     var jsonViewerPreferredMode: JSONViewMode
-    var displaySchemasInSidebar: Bool
 
     static let `default` = EditorSettings(
         showLineNumbers: true,
@@ -86,8 +85,7 @@ struct EditorSettings: Codable, Equatable {
         vimModeEnabled: false,
         uppercaseKeywords: false,
         queryParametersEnabled: true,
-        jsonViewerPreferredMode: .text,
-        displaySchemasInSidebar: false
+        jsonViewerPreferredMode: .text
     )
 
     init(
@@ -98,8 +96,7 @@ struct EditorSettings: Codable, Equatable {
         vimModeEnabled: Bool = false,
         uppercaseKeywords: Bool = false,
         queryParametersEnabled: Bool = true,
-        jsonViewerPreferredMode: JSONViewMode = .text,
-        displaySchemasInSidebar: Bool = false
+        jsonViewerPreferredMode: JSONViewMode = .text
     ) {
         self.showLineNumbers = showLineNumbers
         self.highlightCurrentLine = highlightCurrentLine
@@ -109,7 +106,6 @@ struct EditorSettings: Codable, Equatable {
         self.uppercaseKeywords = uppercaseKeywords
         self.queryParametersEnabled = queryParametersEnabled
         self.jsonViewerPreferredMode = jsonViewerPreferredMode
-        self.displaySchemasInSidebar = displaySchemasInSidebar
     }
 
     init(from decoder: Decoder) throws {
@@ -122,7 +118,6 @@ struct EditorSettings: Codable, Equatable {
         uppercaseKeywords = try container.decodeIfPresent(Bool.self, forKey: .uppercaseKeywords) ?? false
         queryParametersEnabled = try container.decodeIfPresent(Bool.self, forKey: .queryParametersEnabled) ?? true
         jsonViewerPreferredMode = try container.decodeIfPresent(JSONViewMode.self, forKey: .jsonViewerPreferredMode) ?? .text
-        displaySchemasInSidebar = try container.decodeIfPresent(Bool.self, forKey: .displaySchemasInSidebar) ?? false
     }
 
     /// Clamped tab width (1-16)